RW-5381 moving BigQueryService, CohortQueryBuilder and ReviewQueryBuilder into service layer
