+++ api/src/main/java/org/pmiops/workbench/api/CohortReviewController.java
@@ -1,66 +1,29 @@
 package org.pmiops.workbench.api;
 
-import static org.pmiops.workbench.model.FilterColumns.AGE_AT_EVENT;
-import static org.pmiops.workbench.model.FilterColumns.ANSWER;
-import static org.pmiops.workbench.model.FilterColumns.DOMAIN;
-import static org.pmiops.workbench.model.FilterColumns.DOSE;
-import static org.pmiops.workbench.model.FilterColumns.FIRST_MENTION;
-import static org.pmiops.workbench.model.FilterColumns.LAST_MENTION;
-import static org.pmiops.workbench.model.FilterColumns.NUM_MENTIONS;
-import static org.pmiops.workbench.model.FilterColumns.QUESTION;
-import static org.pmiops.workbench.model.FilterColumns.REF_RANGE;
-import static org.pmiops.workbench.model.FilterColumns.ROUTE;
-import static org.pmiops.workbench.model.FilterColumns.SOURCE_CODE;
-import static org.pmiops.workbench.model.FilterColumns.SOURCE_CONCEPT_ID;
-import static org.pmiops.workbench.model.FilterColumns.SOURCE_NAME;
-import static org.pmiops.workbench.model.FilterColumns.SOURCE_VOCABULARY;
-import static org.pmiops.workbench.model.FilterColumns.STANDARD_CODE;
-import static org.pmiops.workbench.model.FilterColumns.STANDARD_CONCEPT_ID;
-import static org.pmiops.workbench.model.FilterColumns.STANDARD_NAME;
-import static org.pmiops.workbench.model.FilterColumns.STANDARD_VOCABULARY;
 import static org.pmiops.workbench.model.FilterColumns.START_DATETIME;
-import static org.pmiops.workbench.model.FilterColumns.STRENGTH;
-import static org.pmiops.workbench.model.FilterColumns.SURVEY_NAME;
-import static org.pmiops.workbench.model.FilterColumns.UNIT;
-import static org.pmiops.workbench.model.FilterColumns.VALUE_AS_NUMBER;
-import static org.pmiops.workbench.model.FilterColumns.VISIT_TYPE;
 
-import com.google.cloud.bigquery.BigQueryException;
-import com.google.cloud.bigquery.FieldValue;
-import com.google.cloud.bigquery.TableResult;
 import com.google.common.collect.ImmutableList;
-import com.google.gson.Gson;
-import java.sql.Date;
 import java.sql.Timestamp;
 import java.time.Clock;
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import javax.inject.Provider;
 import org.pmiops.workbench.cohortbuilder.CohortBuilderService;
-import org.pmiops.workbench.cohortbuilder.CohortQueryBuilder;
-import org.pmiops.workbench.cohortbuilder.ParticipantCriteria;
 import org.pmiops.workbench.cohortreview.CohortReviewService;
-import org.pmiops.workbench.cohortreview.ReviewQueryBuilder;
 import org.pmiops.workbench.cohortreview.util.PageRequest;
 import org.pmiops.workbench.cohortreview.util.ParticipantCohortStatusDbInfo;
 import org.pmiops.workbench.db.dao.UserRecentResourceService;
 import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.db.model.DbParticipantCohortStatus;
-import org.pmiops.workbench.db.model.DbParticipantCohortStatusKey;
-import org.pmiops.workbench.db.model.DbStorageEnums;
 import org.pmiops.workbench.db.model.DbUser;
 import org.pmiops.workbench.exceptions.BadRequestException;
 import org.pmiops.workbench.exceptions.NotFoundException;
-import org.pmiops.workbench.model.CohortChartData;
 import org.pmiops.workbench.model.CohortChartDataListResponse;
 import org.pmiops.workbench.model.CohortReview;
 import org.pmiops.workbench.model.CohortReviewListResponse;
 import org.pmiops.workbench.model.CohortReviewWithCountResponse;
-import org.pmiops.workbench.model.CohortStatus;
 import org.pmiops.workbench.model.CreateReviewRequest;
 import org.pmiops.workbench.model.CriteriaType;
 import org.pmiops.workbench.model.Domain;
@@ -69,18 +32,14 @@ import org.pmiops.workbench.model.FilterColumns;
 import org.pmiops.workbench.model.ModifyCohortStatusRequest;
 import org.pmiops.workbench.model.ModifyParticipantCohortAnnotationRequest;
 import org.pmiops.workbench.model.PageFilterRequest;
-import org.pmiops.workbench.model.ParticipantChartData;
 import org.pmiops.workbench.model.ParticipantChartDataListResponse;
 import org.pmiops.workbench.model.ParticipantCohortAnnotation;
 import org.pmiops.workbench.model.ParticipantCohortAnnotationListResponse;
 import org.pmiops.workbench.model.ParticipantCohortStatus;
-import org.pmiops.workbench.model.ParticipantData;
 import org.pmiops.workbench.model.ParticipantDataCountResponse;
 import org.pmiops.workbench.model.ParticipantDataListResponse;
 import org.pmiops.workbench.model.ReviewStatus;
-import org.pmiops.workbench.model.SearchRequest;
 import org.pmiops.workbench.model.SortOrder;
-import org.pmiops.workbench.model.Vocabulary;
 import org.pmiops.workbench.model.VocabularyListResponse;
 import org.pmiops.workbench.model.WorkspaceAccessLevel;
 import org.pmiops.workbench.workspaces.WorkspaceService;
@@ -106,9 +65,6 @@ public class CohortReviewController implements CohortReviewApiDelegate {
 
   private final CohortBuilderService cohortBuilderService;
   private final CohortReviewService cohortReviewService;
-  private final BigQueryService bigQueryService;
-  private final CohortQueryBuilder cohortQueryBuilder;
-  private final ReviewQueryBuilder reviewQueryBuilder;
   private final UserRecentResourceService userRecentResourceService;
   private final Provider<DbUser> userProvider;
   private final WorkspaceService workspaceService;
@@ -117,19 +73,13 @@ public class CohortReviewController implements CohortReviewApiDelegate {
   @Autowired
   CohortReviewController(
       CohortReviewService cohortReviewService,
-      BigQueryService bigQueryService,
       CohortBuilderService cohortBuilderService,
-      CohortQueryBuilder cohortQueryBuilder,
-      ReviewQueryBuilder reviewQueryBuilder,
       UserRecentResourceService userRecentResourceService,
       Provider<DbUser> userProvider,
       WorkspaceService workspaceService,
       Clock clock) {
     this.cohortReviewService = cohortReviewService;
-    this.bigQueryService = bigQueryService;
     this.cohortBuilderService = cohortBuilderService;
-    this.cohortQueryBuilder = cohortQueryBuilder;
-    this.reviewQueryBuilder = reviewQueryBuilder;
     this.userRecentResourceService = userRecentResourceService;
     this.userProvider = userProvider;
     this.workspaceService = workspaceService;
@@ -163,7 +113,7 @@ public class CohortReviewController implements CohortReviewApiDelegate {
     try {
       cohortReview = cohortReviewService.findCohortReview(cohortId, cdrVersionId);
     } catch (NotFoundException nfe) {
-      cohortReview = initializeCohortReview(cdrVersionId, cohort);
+      cohortReview = cohortReviewService.initializeCohortReview(cdrVersionId, cohort);
       cohortReview = cohortReviewService.saveCohortReview(cohortReview, userProvider.get());
     }
     if (cohortReview.getReviewSize() > 0) {
@@ -173,21 +123,12 @@ public class CohortReviewController implements CohortReviewApiDelegate {
               cohortId, cdrVersionId));
     }
 
-    SearchRequest searchRequest =
-        new Gson().fromJson(getCohortDefinition(cohort), SearchRequest.class);
-
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(
-                cohortQueryBuilder.buildRandomParticipantQuery(
-                    new ParticipantCriteria(searchRequest), request.getSize(), 0L)));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
     List<DbParticipantCohortStatus> participantCohortStatuses =
-        createParticipantCohortStatusesList(cohortReview.getCohortReviewId(), result, rm);
+        cohortReviewService.createDbParticipantCohortStatusesList(
+            cohort, request.getSize(), cohortReview.getCohortReviewId());
 
     cohortReview
-        .reviewSize(Long.valueOf(participantCohortStatuses.size()))
+        .reviewSize((long) participantCohortStatuses.size())
         .reviewStatus(ReviewStatus.CREATED);
 
     // when saving ParticipantCohortStatuses to the database the long value of birthdate is mutated.
@@ -270,33 +211,15 @@ public class CohortReviewController implements CohortReviewApiDelegate {
     }
 
     CohortReview cohortReview = cohortReviewService.findCohortReview(cohortReviewId);
-    DbCohort cohort = cohortReviewService.findCohort(cohortReview.getCohortId());
+    DbCohort dbCohort = cohortReviewService.findCohort(cohortReview.getCohortId());
     workspaceService.getWorkspaceEnforceAccessLevelAndSetCdrVersion(
         workspaceNamespace, workspaceId, WorkspaceAccessLevel.READER);
-
-    SearchRequest searchRequest =
-        new Gson().fromJson(getCohortDefinition(cohort), SearchRequest.class);
-
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(
-                cohortQueryBuilder.buildDomainChartInfoCounterQuery(
-                    new ParticipantCriteria(searchRequest),
-                    Objects.requireNonNull(Domain.fromValue(domain)),
-                    chartLimit)));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
-    CohortChartDataListResponse response = new CohortChartDataListResponse();
-    response.count(cohortReview.getMatchedParticipantCount());
-    for (List<FieldValue> row : result.iterateAll()) {
-      response.addItemsItem(
-          new CohortChartData()
-              .name(bigQueryService.getString(row, rm.get("name")))
-              .conceptId(bigQueryService.getLong(row, rm.get("conceptId")))
-              .count(bigQueryService.getLong(row, rm.get("count"))));
-    }
-
-    return ResponseEntity.ok(response);
+    return ResponseEntity.ok(
+        new CohortChartDataListResponse()
+            .count(cohortReview.getMatchedParticipantCount())
+            .items(
+                cohortReviewService.findCohortChartData(
+                    dbCohort, Objects.requireNonNull(Domain.fromValue(domain)), chartLimit)));
   }
 
   @Override
@@ -330,18 +253,11 @@ public class CohortReviewController implements CohortReviewApiDelegate {
     workspaceService.getWorkspaceEnforceAccessLevelAndSetCdrVersion(
         workspaceNamespace, workspaceId, WorkspaceAccessLevel.READER);
 
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(
-                reviewQueryBuilder.buildChartDataQuery(
+    return ResponseEntity.ok(
+        new ParticipantChartDataListResponse()
+            .items(
+                cohortReviewService.findParticipantChartData(
                     participantId, Objects.requireNonNull(Domain.fromValue(domain)), chartLimit)));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
-    ParticipantChartDataListResponse response = new ParticipantChartDataListResponse();
-    for (List<FieldValue> row : result.iterateAll()) {
-      response.addItemsItem(convertRowToChartData(rm, row));
-    }
-    return ResponseEntity.ok(response);
   }
 
   @Override
@@ -391,7 +307,7 @@ public class CohortReviewController implements CohortReviewApiDelegate {
       participantCohortStatuses =
           cohortReviewService.findAll(cohortReview.getCohortReviewId(), pageRequest);
     } catch (NotFoundException nfe) {
-      cohortReview = initializeCohortReview(cdrVersionId, cohort);
+      cohortReview = cohortReviewService.initializeCohortReview(cdrVersionId, cohort);
     }
 
     cohortReview.participantCohortStatuses(participantCohortStatuses);
@@ -417,17 +333,11 @@ public class CohortReviewController implements CohortReviewApiDelegate {
       PageFilterRequest request) {
     workspaceService.getWorkspaceEnforceAccessLevelAndSetCdrVersion(
         workspaceNamespace, workspaceId, WorkspaceAccessLevel.READER);
-    // get the total records count for this participant per specified domain
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(
-                reviewQueryBuilder.buildCountQuery(
-                    participantId, request.getDomain(), createPageRequest(request))));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-    ParticipantDataCountResponse response =
+    return ResponseEntity.ok(
         new ParticipantDataCountResponse()
-            .count(bigQueryService.getLong(result.iterateAll().iterator().next(), rm.get("count")));
-    return ResponseEntity.ok(response);
+            .count(
+                cohortReviewService.findParticipantCount(
+                    participantId, request.getDomain(), createPageRequest(request))));
   }
 
   @Override
@@ -443,19 +353,11 @@ public class CohortReviewController implements CohortReviewApiDelegate {
     // this validates that the participant is in the requested review.
     cohortReviewService.findParticipantCohortStatus(cohortReviewId, participantId);
 
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(
-                reviewQueryBuilder.buildQuery(
+    return ResponseEntity.ok(
+        new ParticipantDataListResponse()
+            .items(
+                cohortReviewService.findParticipantData(
                     participantId, request.getDomain(), createPageRequest(request))));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
-    ParticipantDataListResponse response = new ParticipantDataListResponse();
-    for (List<FieldValue> row : result.iterateAll()) {
-      response.addItemsItem(convertRowToParticipantData(rm, row, request.getDomain()));
-    }
-
-    return ResponseEntity.ok(response);
   }
 
   @Override
@@ -464,20 +366,8 @@ public class CohortReviewController implements CohortReviewApiDelegate {
     workspaceService.getWorkspaceEnforceAccessLevelAndSetCdrVersion(
         workspaceNamespace, workspaceId, WorkspaceAccessLevel.READER);
 
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(reviewQueryBuilder.buildVocabularyDataQuery()));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-
-    VocabularyListResponse response = new VocabularyListResponse();
-    for (List<FieldValue> row : result.iterateAll()) {
-      response.addItemsItem(
-          new Vocabulary()
-              .domain(bigQueryService.getString(row, rm.get("domain")))
-              .type(bigQueryService.getString(row, rm.get("type")))
-              .vocabulary(bigQueryService.getString(row, rm.get("vocabulary"))));
-    }
-    return ResponseEntity.ok(response);
+    return ResponseEntity.ok(
+        new VocabularyListResponse().items(cohortReviewService.findVocabularies()));
   }
 
   @Override
@@ -529,81 +419,6 @@ public class CohortReviewController implements CohortReviewApiDelegate {
             new Timestamp(clock.instant().toEpochMilli())));
   }
 
-  /** Helper method to create a new {@link CohortReview}. */
-  private CohortReview initializeCohortReview(Long cdrVersionId, DbCohort cohort) {
-    SearchRequest request = new Gson().fromJson(getCohortDefinition(cohort), SearchRequest.class);
-
-    TableResult result =
-        bigQueryService.executeQuery(
-            bigQueryService.filterBigQueryConfig(
-                cohortQueryBuilder.buildParticipantCounterQuery(new ParticipantCriteria(request))));
-    Map<String, Integer> rm = bigQueryService.getResultMapper(result);
-    List<FieldValue> row = result.iterateAll().iterator().next();
-    long cohortCount = bigQueryService.getLong(row, rm.get("count"));
-
-    return createNewCohortReview(cohort, cdrVersionId, cohortCount);
-  }
-
-  /**
-   * Helper method that builds a list of {@link DbParticipantCohortStatus} from BigQuery results.
-   */
-  private List<DbParticipantCohortStatus> createParticipantCohortStatusesList(
-      Long cohortReviewId, TableResult result, Map<String, Integer> rm) {
-    List<DbParticipantCohortStatus> participantCohortStatuses = new ArrayList<>();
-    for (List<FieldValue> row : result.iterateAll()) {
-      String birthDateTimeString = bigQueryService.getString(row, rm.get("birth_datetime"));
-      if (birthDateTimeString == null) {
-        throw new BigQueryException(
-            500, "birth_datetime is null at position: " + rm.get("birth_datetime"));
-      }
-      java.util.Date birthDate =
-          Date.from(Instant.ofEpochMilli(Double.valueOf(birthDateTimeString).longValue() * 1000));
-      participantCohortStatuses.add(
-          new DbParticipantCohortStatus()
-              .participantKey(
-                  new DbParticipantCohortStatusKey(
-                      cohortReviewId, bigQueryService.getLong(row, rm.get("person_id"))))
-              .status(DbStorageEnums.cohortStatusToStorage(CohortStatus.NOT_REVIEWED))
-              .birthDate(new Date(birthDate.getTime()))
-              .genderConceptId(bigQueryService.getLong(row, rm.get("gender_concept_id")))
-              .raceConceptId(bigQueryService.getLong(row, rm.get("race_concept_id")))
-              .ethnicityConceptId(bigQueryService.getLong(row, rm.get("ethnicity_concept_id")))
-              .sexAtBirthConceptId(bigQueryService.getLong(row, rm.get("sex_at_birth_concept_id")))
-              .deceased(bigQueryService.getBoolean(row, rm.get("deceased"))));
-    }
-    return participantCohortStatuses;
-  }
-
-  /**
-   * Helper to method that consolidates access to Cohort Definition. Will throw a {@link
-   * NotFoundException} if {@link DbCohort#getCriteria()} return null.
-   */
-  private String getCohortDefinition(DbCohort cohort) {
-    String definition = cohort.getCriteria();
-    if (definition == null) {
-      throw new NotFoundException(
-          String.format(
-              "Not Found: No Cohort definition matching cohortId: %s", cohort.getCohortId()));
-    }
-    return definition;
-  }
-
-  /** Helper method that constructs a {@link CohortReview} with the specified ids and count. */
-  private CohortReview createNewCohortReview(DbCohort cohort, Long cdrVersionId, Long cohortCount) {
-    return new CohortReview()
-        .cohortId(cohort.getCohortId())
-        .cohortDefinition(getCohortDefinition(cohort))
-        .cohortName(cohort.getName())
-        .description(cohort.getDescription())
-        .cdrVersionId(cdrVersionId)
-        .matchedParticipantCount(cohortCount)
-        .creationTime(new Timestamp(clock.instant().toEpochMilli()).getTime())
-        .lastModifiedTime(new Timestamp(clock.instant().toEpochMilli()).getTime())
-        .reviewedCount(0L)
-        .reviewSize(0L)
-        .reviewStatus(ReviewStatus.NONE);
-  }
-
   /** Helper method that converts sortOrder if gender, race or ethnicity. */
   private void convertGenderRaceEthnicitySortOrder(PageRequest pageRequest) {
     String sortColumn = pageRequest.getSortColumn();
@@ -631,9 +446,8 @@ public class CohortReviewController implements CohortReviewApiDelegate {
   private PageRequest createPageRequest(PageFilterRequest request) {
     FilterColumns col = request.getDomain() == null ? FilterColumns.PARTICIPANTID : START_DATETIME;
     String sortColumn = Optional.ofNullable(request.getSortColumn()).orElse(col).toString();
-    int pageParam = Optional.ofNullable(request.getPage()).orElse(CohortReviewController.PAGE);
-    int pageSizeParam =
-        Optional.ofNullable(request.getPageSize()).orElse(CohortReviewController.PAGE_SIZE);
+    int pageParam = Optional.ofNullable(request.getPage()).orElse(PAGE);
+    int pageSizeParam = Optional.ofNullable(request.getPageSize()).orElse(PAGE_SIZE);
     SortOrder sortOrderParam = Optional.ofNullable(request.getSortOrder()).orElse(SortOrder.ASC);
     return new PageRequest()
         .page(pageParam)
@@ -642,57 +456,5 @@ public class CohortReviewController implements CohortReviewApiDelegate {
         .sortColumn(sortColumn)
         .filters(
             request.getFilters() == null ? new ArrayList<>() : request.getFilters().getItems());
-  }
-
-  /** Helper method to convert a collection of {@link FieldValue} to {@link ParticipantData}. */
-  private ParticipantData convertRowToParticipantData(
-      Map<String, Integer> rm, List<FieldValue> row, Domain domain) {
-    if (!domain.equals(Domain.SURVEY)) {
-      return new ParticipantData()
-          .itemDate(bigQueryService.getDateTime(row, rm.get(START_DATETIME.toString())))
-          .domain(bigQueryService.getString(row, rm.get(DOMAIN.toString())))
-          .standardName(bigQueryService.getString(row, rm.get(STANDARD_NAME.toString())))
-          .ageAtEvent(bigQueryService.getLong(row, rm.get(AGE_AT_EVENT.toString())).intValue())
-          .standardConceptId(bigQueryService.getLong(row, rm.get(STANDARD_CONCEPT_ID.toString())))
-          .sourceConceptId(bigQueryService.getLong(row, rm.get(SOURCE_CONCEPT_ID.toString())))
-          .standardVocabulary(
-              bigQueryService.getString(row, rm.get(STANDARD_VOCABULARY.toString())))
-          .sourceVocabulary(bigQueryService.getString(row, rm.get(SOURCE_VOCABULARY.toString())))
-          .sourceName(bigQueryService.getString(row, rm.get(SOURCE_NAME.toString())))
-          .sourceCode(bigQueryService.getString(row, rm.get(SOURCE_CODE.toString())))
-          .standardCode(bigQueryService.getString(row, rm.get(STANDARD_CODE.toString())))
-          .value(bigQueryService.getString(row, rm.get(VALUE_AS_NUMBER.toString())))
-          .visitType(bigQueryService.getString(row, rm.get(VISIT_TYPE.toString())))
-          .numMentions(bigQueryService.getString(row, rm.get(NUM_MENTIONS.toString())))
-          .firstMention(
-              row.get(rm.get(FIRST_MENTION.toString())).isNull()
-                  ? ""
-                  : bigQueryService.getDateTime(row, rm.get(FIRST_MENTION.toString())))
-          .lastMention(
-              row.get(rm.get(LAST_MENTION.toString())).isNull()
-                  ? ""
-                  : bigQueryService.getDateTime(row, rm.get(LAST_MENTION.toString())))
-          .unit(bigQueryService.getString(row, rm.get(UNIT.toString())))
-          .dose(bigQueryService.getString(row, rm.get(DOSE.toString())))
-          .strength(bigQueryService.getString(row, rm.get(STRENGTH.toString())))
-          .route(bigQueryService.getString(row, rm.get(ROUTE.toString())))
-          .refRange(bigQueryService.getString(row, rm.get(REF_RANGE.toString())));
-    } else {
-      return new ParticipantData()
-          .itemDate(bigQueryService.getDateTime(row, rm.get(START_DATETIME.toString())))
-          .survey(bigQueryService.getString(row, rm.get(SURVEY_NAME.toString())))
-          .question(bigQueryService.getString(row, rm.get(QUESTION.toString())))
-          .answer(bigQueryService.getString(row, rm.get(ANSWER.toString())));
-    }
-  }
-
-  private ParticipantChartData convertRowToChartData(
-      Map<String, Integer> rm, List<FieldValue> row) {
-    return new ParticipantChartData()
-        .standardName(bigQueryService.getString(row, rm.get("standardName")))
-        .standardVocabulary(bigQueryService.getString(row, rm.get("standardVocabulary")))
-        .startDate(bigQueryService.getDate(row, rm.get("startDate")))
-        .ageAtEvent(bigQueryService.getLong(row, rm.get("ageAtEvent")).intValue())
-        .rank(bigQueryService.getLong(row, rm.get("rank")).intValue());
   }
 }
