+++ api/src/main/java/org/pmiops/workbench/cohortreview/CohortReviewService.java
@@ -8,11 +8,16 @@ import org.pmiops.workbench.db.model.DbCohortReview;
 import org.pmiops.workbench.db.model.DbParticipantCohortAnnotation;
 import org.pmiops.workbench.db.model.DbParticipantCohortStatus;
 import org.pmiops.workbench.db.model.DbUser;
+import org.pmiops.workbench.model.CohortChartData;
 import org.pmiops.workbench.model.CohortReview;
 import org.pmiops.workbench.model.CohortStatus;
+import org.pmiops.workbench.model.Domain;
 import org.pmiops.workbench.model.ModifyParticipantCohortAnnotationRequest;
+import org.pmiops.workbench.model.ParticipantChartData;
 import org.pmiops.workbench.model.ParticipantCohortAnnotation;
 import org.pmiops.workbench.model.ParticipantCohortStatus;
+import org.pmiops.workbench.model.ParticipantData;
+import org.pmiops.workbench.model.Vocabulary;
 
 public interface CohortReviewService {
 
@@ -88,4 +93,30 @@ public interface CohortReviewService {
    */
   List<ParticipantCohortAnnotation> findParticipantCohortAnnotations(
       Long cohortReviewId, Long participantId);
+
+  /** Initialize a {@link CohortReview} */
+  CohortReview initializeCohortReview(Long cdrVersionId, DbCohort dbCohort);
+
+  /**
+   * Create a list of {@link DbParticipantCohortStatus} for the specified cohort, requestSize and
+   * review id.
+   */
+  List<DbParticipantCohortStatus> createDbParticipantCohortStatusesList(
+      DbCohort dbCohort, Integer requestSize, Long cohortReviewId);
+
+  /** Find a list of {@link CohortChartData} for the specified cohort and domain. */
+  List<CohortChartData> findCohortChartData(DbCohort dbCohort, Domain domain, int limit);
+
+  /** Find a list of {@link ParticipantChartData} for the specified participant id and domain. */
+  List<ParticipantChartData> findParticipantChartData(Long participantId, Domain domain, int limit);
+
+  /** Find participant count. */
+  Long findParticipantCount(Long participantId, Domain domain, PageRequest pageRequest);
+
+  /** Find list of {@link ParticipantData} for specifice participant id and domain. */
+  List<ParticipantData> findParticipantData(
+      Long participantId, Domain domain, PageRequest pageRequest);
+
+  /** Find list of {@link Vocabulary}.* */
+  List<Vocabulary> findVocabularies();
 }
